Adding a folder with extra metadata
This should be a provisional folder. But at the same time, I am not sure if I will ever change this.
The problem with trying to extract the data automatically, it just takes for freaking forever and I am not sure
what the practicality, as the metadata are not really big, so it can easily be stored on github