Solving a dumb mistake in the data formatting
There is no point in solving this automatically, as it is an unexplicable formating of decimal numbers in the data. There might be better ways to solve it but since it only affects annotations (smaller file) in one specific dataset, I will just add it to the github repository and let people import it all together. I guess this could be a solution in similar problems in the future